managed_context_ops: carve pressure.rs (pure move)
Context-usage snapshots and the pressure machinery: snapshot
emit/refresh, preflight rewind-only/density gates, tool allow-lists,
dashboard-command classification, recovery and density kickstart/
handoff texts, surgical recovery, and follow-up replay text; 34 tests
and the surgical catalog fixture ride along.
